refactor(frontend): cleanup code ss-192
