+++ apps/frontend/src/libs/components/select/libs/constants/constants.ts
@@ -1,0 +1,1 @@
+export { selectClassNames } from "./select-class-names.constant.js";
