+++ apps/frontend/src/libs/components/select/libs/constants/select-class-names.constant.ts
@@ -1,0 +1,27 @@
+import { combineClassNames } from "~/libs/helpers/helpers.js";
+
+import styles from "../../styles.module.css";
+
+const selectClassNames = {
+	control: (): string => styles["control"] as string,
+	menu: (): string => styles["menu"] as string,
+	multiValue: (): string => styles["multi-value"] as string,
+	multiValueLabel: (): string => styles["multi-value-label"] as string,
+	multiValueRemove: (): string => styles["multi-value-remove"] as string,
+	option: ({
+		isFocused,
+		isSelected,
+	}: {
+		isFocused: boolean;
+		isSelected: boolean;
+	}): string =>
+		combineClassNames(
+			styles["option"],
+			isFocused && styles["option-focused"],
+			isSelected && styles["option-selected"],
+		),
+	placeholder: (): string => styles["placeholder"] as string,
+	singleValue: (): string => styles["single-value"] as string,
+};
+
+export { selectClassNames };
