+++ apps/frontend/src/libs/components/select/libs/mappers/mappers.ts
@@ -1,0 +1,1 @@
+export { mapSelectValue } from "./select-value.mapper.js";
